fix home page

+++ controllers/tryout.js
@@ -53,8 +53,10 @@ module.exports.updateTryout = async (req, res) => {
 module.exports.deleteTryout = async (req, res) => {
     const {tryoutId} = req.params
     const tryout = await Tryout.findById(tryoutId)
-    for (const question of tryout.question) {
-        await Question.findByIdAndDelete(question._id)
+    if(tryout.question) {
+        for (const question of tryout.question) {
+            await Question.findByIdAndDelete(question._id)
+        }
     }
     await Tryout.findByIdAndDelete(tryoutId)
     res.redirect('/tryout')
